Fix READMEs fetched from GitHub
The previous technique for fetching GitHub READMEs was a bit broken because it relied on the repo’s default branch being `main`, which turned out not to always be the case.

This commit uses the GitHub REST API to fetch the READMEs instead. This requires a GitHub Access Token to be added to the build environment variables as `PUBLIC_GITHUB_TOKEN`, but is a much more reliable way to get the content.

Once I’d done that, I realised a bunch more rendering limitations with just plonking the raw Markdown from GitHub on a page: broken images & links and rendering anomalies. So I’ve built a remark pipeline to try to quickly emulate GitHub’s rendering and fix the images and links. It’s a bit like npm’s `marky-markdown`, but that doesn’t seem to be compatible with modern projects.